New tests, fixes and overall speed improvement
Test app was moved to this project from a separate one because it's a pain in the ass to wait for jitpack to build the library when changing something in it
Greatly improved speed of all the SAF operations by introducing CachingDocumentFile which caches the result of a SAF operation (like getName()) and then returns this cached value instead of doing an IPC every single time
Introduced directory snapshots. What it does is it loads every single file inside a directory (that you choose yourself), preloads all the needed information (like name or length), stores it in SnapshotDocumentFile (which is even faster than CachingDocumentFile because it uses the preloaded info instead of doing IPCs) and doing it in a single IPC. Then it stores all of that info instead FastFileSearchTree. When the snapshot is not needed anymore, all the preloaded information get deleted from the FastFileSearchTree. It's best to use snapshots when there is a directory with multiple files (or even with subdirectories) and you need to check whether multiple files exist in that directory.